rework the EC key generation mechanism
Before this commit, parameters of supported EC curves were saved in the
applet. This is not feasible when trying to support a large number of
curves because the applet has size limitations.
The new version requires the host to send the EC parameters. This has
the additional advantage that (depending on the host
middleware/driver/tool) new curves can be supported without chaning the
API.