Add support for EC curves with small parameters
If a parameter has fewer non-zero bytes than the field length, Java Card
returns the parameter with the smaller length. This is unexpected, because when
setting the parameter, it has to be filled up with zeroes (no length may be
specified). This change allows curves like secp192k1 or secp256k1 to be used
with the applet.