Implement ERASE_CARD apdu and introduce SO PIN
Smartcard can now also be initialised using

pkcs11-tool --init-token ...
pkcs11-tool --init-pin ...

A separate SO PIN (with reference 0x0F) was introduced and has to be provided during token initialisation.
PUK is now used only for user PIN unblocking and is initially set (when using C_InitToken()) to SO PIN value.
However SO PIN can later be changed if needed.

The bulk of changes was made to processVerify() and processChangeReferenceData() functions.

ERASE_CARD apdu resets token state to STATE_CREATION including all the keys and pkcs#15 fs. As there is a possibilty that incomplete initialisation left the fs in an inconsistent state
an INITIALISE apdu was necessary. It just clears the FS if the token is in STATE_CREATION state and possibly reconfigures card with new default values.

As some (nontrivial) changes were required also on OpenSC part, applet version number was increased to 0.7.